refacto: rewrite CPOM altimetry tests
